feat(pagination) 2 endpoints rewritten for 1, changed registration logic, changed course markup

+++ bot/api/api_models.py
@@ -22,7 +22,7 @@ class UserBaseDto(BaseModel):
     last_name: str
     email: str
     phone_number: str
-    time_zone: int
+    time_zone: float
 
 
 class UserDto(UserBaseDto):
